Hide Columns
When this VBA script is executed (for example by clicking a form that is linked to this script), all columns with lowercase first letters will be hidden. Columns with Capital Letters will remain. 

This script requires a region formatted as a table. This script also only works on the table. All other columns will not be touched.